+++ visualization/app/codeCharta/state/store/appSettings/invertArea/invertArea.reducer.ts
@@ -1,9 +1,11 @@
-import { InvertAreaAction, InvertAreaActions, setInvertArea } from "./invertArea.actions"
+import { InvertAreaAction, InvertAreaActions, setInvertArea, ToggleInvertingAreaAction } from "./invertArea.actions"
 
-export function invertArea(state = setInvertArea().payload, action: InvertAreaAction) {
+export function invertArea(state = setInvertArea().payload, action: InvertAreaAction | ToggleInvertingAreaAction) {
 	switch (action.type) {
 		case InvertAreaActions.SET_INVERT_AREA:
 			return action.payload
+		case InvertAreaActions.TOGGLE_INVERTING_AREA:
+			return !state
 		default:
 			return state
 	}