+++ visualization/app/codeCharta/state/store/appSettings/invertArea/invertArea.selector.ts
@@ -1,0 +1,3 @@
+import { CcState } from "../../store"
+
+export const invertAreaSelector = (state: CcState): boolean => state.appSettings.invertArea