+++ visualization/app/codeCharta/state/store/appSettings/isEdgeMetricVisible/isEdgeMetricVisible.actions.ts
@@ -1,0 +1,19 @@
+import { Action } from "redux"
+
+export enum IsEdgeMetricVisibleActions {
+	SET_IS_EDGE_METRIC_VISIBLE = "SET_IS_EDGE_METRIC_VISIBLE"
+}
+
+export interface SetIsEdgeMetricVisibleAction extends Action {
+	type: IsEdgeMetricVisibleActions.SET_IS_EDGE_METRIC_VISIBLE
+}
+
+export type IsEdgeMetricVisibleAction = SetIsEdgeMetricVisibleAction
+
+export function setIsEdgeMetricVisible(): SetIsEdgeMetricVisibleAction {
+	return {
+		type: IsEdgeMetricVisibleActions.SET_IS_EDGE_METRIC_VISIBLE
+	}
+}
+
+export const defaultIsEdgeMetricVisible = true