+++ visualization/app/codeCharta/state/store/appSettings/isLoadingMap/isLoadingMap.reducer.ts
@@ -13,8 +13,8 @@ import { PanelSelectionActions } from "../panelSelection/panelSelection.actions"
 import { SearchPanelModeActions } from "../searchPanelMode/searchPanelMode.actions"
 import { SortingOrderAscendingActions } from "../sortingOrderAscending/sortingOrderAscending.actions"
 import { IsLoadingMapAction, IsLoadingMapActions, setIsLoadingMap } from "./isLoadingMap.actions"
-import { IdToNodeActions } from "../../lookUp/idToNode/idToNode.actions"
-import { IdToBuildingActions } from "../../lookUp/idToBuilding/idToBuilding.actions"
+import { RightClickedNodeDataActions } from "../../appStatus/rightClickedNodeData/rightClickedNodeData.actions"
+import { IsEdgeMetricVisibleActions } from "../isEdgeMetricVisible/isEdgeMetricVisible.actions"
 
 // Todo state actions explicit instead of excluding all others; refs #1547
 const actionsToExclude = [
@@ -27,13 +27,13 @@ const actionsToExclude = [
 	PanelSelectionActions,
 	PresentationModeActions,
 	ExperimentalFeaturesEnabledActions,
+	IsEdgeMetricVisibleActions,
 	ScreenshotToClipboardEnabledActions,
 	HoveredBuildingPathActions,
+	RightClickedNodeDataActions,
 	FocusedNodePathActions,
 	CameraTargetActions,
-	CameraActions,
-	IdToNodeActions,
-	IdToBuildingActions
+	CameraActions
 ]
 
 export function isLoadingMap(state = setIsLoadingMap().payload, action: IsLoadingMapAction) {