+++ visualization/app/codeCharta/state/store/lookUp/idToBuilding/idToBuilding.selector.ts
@@ -1,3 +1,0 @@
-import { CcState } from "../../store"
-
-export const idToBuildingSelector = (state: CcState) => state.lookUp.idToBuilding