+++ visualization/app/codeCharta/state/store/lookUp/idToNode/idToNode.splitter.ts
@@ -1,6 +1,0 @@
-import { setIdToNode } from "./idToNode.actions"
-import { CodeMapNode } from "../../../../codeCharta.model"
-
-export function splitIdToNodeAction(payload: Map<number, CodeMapNode>) {
-	return setIdToNode(payload)
-}