+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.e2e.ts
@@ -2,11 +2,13 @@ import { goto } from "../../../puppeteer.helper"
 import { RibbonBarPageObject } from "./ribbonBar.po"
 import { SearchPanelPageObject } from "../searchPanel/searchPanel.po"
 import { AreaSettingsPanelPageObject } from "../areaSettingsPanel/areaSettingsPanel.po"
+import { FileChooserPageObject } from "../fileChooser/fileChooser.po"
 
 //Commented out flaky test
 
 describe("RibbonBar", () => {
 	let searchPanel: SearchPanelPageObject
+	let fileChooser: FileChooserPageObject
 	//let searchPanelModeSelector: SearchPanelModeSelectorPageObject
 	let ribbonBar: RibbonBarPageObject
 	//let metricChooser: MetricChooserPageObject
@@ -14,6 +16,7 @@ describe("RibbonBar", () => {
 
 	beforeEach(async () => {
 		searchPanel = new SearchPanelPageObject()
+		fileChooser = new FileChooserPageObject()
 		//searchPanelModeSelector = new SearchPanelModeSelectorPageObject()
 		ribbonBar = new RibbonBarPageObject()
 		//metricChooser = new MetricChooserPageObject()
@@ -47,7 +50,7 @@ describe("RibbonBar", () => {
 	})
 
 	describe("opening and closing ribbon-bar cards", () => {
-		it("searchPanel", async () => {
+		it("search-panel-card", async () => {
 			let isSearchPanelOpen = await searchPanel.toggle()
 			expect(isSearchPanelOpen).toBeTruthy()
 
@@ -55,7 +58,7 @@ describe("RibbonBar", () => {
 			expect(isSearchPanelOpen).toBeFalsy()
 		})
 
-		it("height-metric cad", async () => {
+		it("height-metric card", async () => {
 			const panel = "height-metric"
 
 			let isHeightSettingsPanelOpen = await ribbonBar.togglePanel(panel)
@@ -85,7 +88,7 @@ describe("RibbonBar", () => {
 			expect(isColorSettingsPanelOpen).toBeFalsy()
 		})
 
-		it("edge-metric", async () => {
+		it("edge-metric card", async () => {
 			const panel = "edge-metric"
 
 			let isEdgeSettingsPanelOpen = await ribbonBar.togglePanel(panel)
@@ -124,5 +127,19 @@ describe("RibbonBar", () => {
 		expect(await AreaSettingsPanelPageObject.toggleDefaultMargin()).toBeFalsy()
 
 		expect(await ribbonBar.isPanelOpen(areaPanel)).toBeTruthy()
+	})
+
+	it("should not find edge-metric-panel when file has no edge metrics", async () => {
+		const edgePanel = "edge-metric"
+		await fileChooser.openFiles(["./app/codeCharta/assets/sample3.cc.json"])
+
+		expect(await ribbonBar.isElementPresent(edgePanel)).toBeFalsy()
+	})
+
+	it("should find edge-metric-panel when file has edge metrics", async () => {
+		const edgePanel = "edge-metric"
+		await fileChooser.openFiles(["./app/codeCharta/assets/sample1.cc.json"])
+
+		expect(await ribbonBar.isElementPresent(edgePanel)).toBeTruthy()
 	})
 })