+++ visualization/app/codeCharta/util/usageDataTracker.ts
@@ -1,9 +1,7 @@
 "use strict"
 
 import { getVisibleFileStates, isSingleState } from "../model/files/files.helper"
-import { CodeChartaStorage } from "./codeChartaStorage"
-import { CodeMapNode, NodeType, State } from "../codeCharta.model"
-import { isStandalone } from "./envDetector"
+import { CodeMapNode, NodeType } from "../codeCharta.model"
 import { isActionOfType } from "./reduxHelper"
 import { AreaMetricActions } from "../state/store/dynamicSettings/areaMetric/areaMetric.actions"
 import { HeightMetricActions } from "../state/store/dynamicSettings/heightMetric/heightMetric.actions"
@@ -15,8 +13,8 @@ import { APIVersions } from "../codeCharta.api.model"
 import { getAsApiVersion } from "./fileValidator"
 import { hierarchy } from "d3-hierarchy"
 import { getMedian, pushSorted } from "./nodeDecorator"
-import { RangeSliderController } from "../ui/rangeSlider/rangeSlider.component"
 import { ColorRangeActions } from "../state/store/dynamicSettings/colorRange/colorRange.actions"
+import { FileState } from "../model/files/files"
 
 interface MetaDataTrackingItem {
 	mapId: string
@@ -56,10 +54,12 @@ interface MetricStatistics {
 	metricValues: number[]
 }
 
-function isTrackingAllowed(state: State) {
-	const singleFileStates = getVisibleFileStates(state.files)
+export const TRACKING_DATA_LOCAL_STORAGE_ELEMENT = "CodeCharta::usageData"
 
-	if (!isStandalone() || !isSingleState(state.files) || singleFileStates.length > 1) {
+function isTrackingAllowed(files: FileState[]) {
+	const singleFileStates = getVisibleFileStates(files)
+
+	if (!isSingleState(files) || singleFileStates.length > 1) {
 		return false
 	}
 
@@ -72,12 +72,12 @@ function isTrackingAllowed(state: State) {
 	)
 }
 
-export function trackMapMetaData(state: State) {
-	if (!isTrackingAllowed(state)) {
+export function trackMapMetaData(files: FileState[]) {
+	if (!isTrackingAllowed(files)) {
 		return
 	}
 
-	const singleFileStates = getVisibleFileStates(state.files)
+	const singleFileStates = getVisibleFileStates(files)
 	const fileNodes: CodeMapNode[] = getFileNodes(singleFileStates[0].file.map)
 	const fileMeta = singleFileStates[0].file.fileMeta
 
@@ -90,12 +90,11 @@ export function trackMapMetaData(state: State) {
 		repoCreationDate: fileMeta.repoCreationDate
 	}
 
-	const fileStorage = new CodeChartaStorage()
 	// Make sure that only files within usageData can be read
 	const fileChecksum = trackingDataItem.mapId.replace(/\//g, "")
 
 	try {
-		fileStorage.setItem(`usageData/${fileChecksum}-meta`, JSON.stringify(trackingDataItem))
+		localStorage.setItem(`${TRACKING_DATA_LOCAL_STORAGE_ELEMENT}/${fileChecksum}-meta`, JSON.stringify(trackingDataItem))
 	} catch {
 		// ignore it
 	}
@@ -269,21 +268,10 @@ interface EventTrackingItem {
 	payload: SettingChangedEventPayload | NodeInteractionEventPayload
 }
 
-export function trackEventUsageData(actionType: string, state: State, payload?: any) {
-	if (
-		!isTrackingAllowed(state) ||
-		(!isActionOfType(actionType, AreaMetricActions) &&
-			!isActionOfType(actionType, HeightMetricActions) &&
-			!isActionOfType(actionType, ColorMetricActions) &&
-			!isActionOfType(actionType, ColorRangeActions) &&
-			!isActionOfType(actionType, BlacklistActions) &&
-			!isActionOfType(actionType, FocusedNodePathActions) &&
-			![RangeSliderController.COLOR_RANGE_FROM_UPDATED, RangeSliderController.COLOR_RANGE_TO_UPDATED].includes(actionType))
-	) {
-		return
-	}
+export function trackEventUsageData(actionType: string, files: FileState[], payload?: any) {
+	if (!isTrackingAllowed(files)) return
 
-	const singleFileStates = getVisibleFileStates(state.files)
+	const singleFileStates = getVisibleFileStates(files)
 	const fileMeta = singleFileStates[0].file.fileMeta
 
 	const eventTrackingItem = buildEventTrackingItem(fileMeta.fileChecksum, actionType, payload)
@@ -292,13 +280,11 @@ export function trackEventUsageData(actionType: string, state: State, payload?: 
 	}
 
 	// Make sure that only files within usageData can be read
-	const fileChecksum = getVisibleFileStates(state.files)[0].file.fileMeta.fileChecksum.replace(/\//g, "")
-
-	const fileStorage = new CodeChartaStorage()
+	const fileChecksum = getVisibleFileStates(files)[0].file.fileMeta.fileChecksum.replace(/\//g, "")
 
 	let appendedEvents = ""
 	try {
-		appendedEvents = fileStorage.getItem(`usageData/${fileChecksum}-events`)
+		appendedEvents = localStorage.getItem(`${TRACKING_DATA_LOCAL_STORAGE_ELEMENT}/${fileChecksum}-events`)
 	} catch {
 		// ignore, it no events item exists
 	}
@@ -307,7 +293,10 @@ export function trackEventUsageData(actionType: string, state: State, payload?: 
 		if (appendedEvents.length > 0) {
 			appendedEvents += "\n"
 		}
-		fileStorage.setItem(`usageData/${fileChecksum}-events`, appendedEvents + JSON.stringify(eventTrackingItem))
+		localStorage.setItem(
+			`${TRACKING_DATA_LOCAL_STORAGE_ELEMENT}/${fileChecksum}-events`,
+			appendedEvents + JSON.stringify(eventTrackingItem)
+		)
 	} catch {
 		// ignore tracking errors
 	}